Buffer samples before analyzing them
perf may give us samples before the MMAPs which define the DSOs they
refer to. This leads to breakage when trying to unwind those same
samples. By buffering the samples until the MMAPs have been processed
we can avoid this. The ring buffer the kernel uses for recording is
usually about 512k in size. Thus, by buffering more than that we can
be pretty sure that by the time we start analyzing a sample we have
processed all MMAPs that were in the ring buffer at the time we
received it.

Change-Id: Iec02ff968d82698b9952375a63b915aed54fc15d
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>