Keep data from earlier stack snapshots as fallback for newer ones
The bottom parts of the stack (main(), processEvents(), etc.) are not
very likely to change when the stack is very deep. So, if the stack
snapshot we receive with a sample is too shallow to unwind all of it,
we fall back to data read earlier in order to guess the call chain.

We also mark the number of frames guessed like this in the result.

Change-Id: I0f06e3c87d6ab86021b48444fe02c1c3bc84e203
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>