Make the PID part of the location information
The same combination of address/file/line/column might show up in
different processes, especially if we only know the address and the
other values are defaults. The problem is then that the locations might
be in different binaries and belong to different symbols. Thus, in
order to be able to key the symbols by locationId, we have to make the
PID part of the location.

This prevents the unification of locations from multiple invocations of
the same binary, but as the kernel is free to change the address layout
on each invocation anyway, this should be fine.

Change-Id: If3040a8aba63ae82573063c7fff26533cc77d3d8
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>