Optimize: Only send strings once over the wire
The new StringDefinition is used to intern string data. Thus, commonly
shared strings like file or binary names are only send once, instead
of multiple times for every unique symbol or location definition.

On the client side, this also leads to an easy optimization, as the
strings get interned for free and Qt's implicit sharing will lead
to a reduction in memory consumption.

Change-Id: Icf15d81b9ebe979d00b35db299f3b38a50f0836a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>