Fix elf registering once and for all
When overwriting an mmap, the remaining fragments of the mmap being
overwritten are added as new mmaps, with start, length and offset
adapted. To complement this we use the "pgoff" we get from perf in
order to identify files which are mapped at some offset and avoid
reporting those to dwfl, as dwfl cannot handle them.

In turn, we drop the filtering for PROT_EXEC. This did not work as
JIT compilers can map their memory without PROT_EXEC and later call
mprotect() to make it executable. mprotect is not visible from the
trace.

Change-Id: I2f7525419cd66a0852838b66f8392f0bd94a6af7
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>