Resolve kernel addresses to symbols using /proc/kallsyms mapping
Doing that ourselves is much easier than getting libdwfl to do it
for us as far as I can see. Most notably, it is trivial to parse
the mapping from any path that the user specifies. This will be
required to support mappings from other machines.

The benchmark shows that this is takes some time, but it's not
too bad. On my machine, parsing the file takes about ~60ms. This
could be optimized somewhat to around ~40ms by using raw C API,
but it will probably be less safe then. I decided against that,
since the file will only be parsed once.

Change-Id: I8ef53406fcd295676f07618676a8679bdd13b621
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>