Send global perf attributes
So far, only record-specific attributes were send to the client.
This meant one could not investigate the attributes of a a perf.data
file created by `perf record` and stored to disk directly.

This patch fixes that and will now also send global attributes
stored directly within the perf header.

Change-Id: I3def876b0c61a9366a50dd8c74272394dc24929f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>