Publish lost record events
Previously these events where stored in a list but never actually
used for anything. Now, the record data is published to the output
data stream, i.e. only the fact that something got lost, and when.

This information is enough to enable clients to show appropriate
warning messages that the data is not to be fully trusted, like
`perf report` does when it talks about "lost chunks" which
is the total sum of all lost records.

Change-Id: I7a923062b70a6e5e1ac5ac1600b70d9703fc4032
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>