Fix test when kallsyms mapping is disabled/broken
On our CI, the kallsyms file only contains 0 addresses, i.e. lines
of the form

...
0000000000000000 b current_type [floppy]
0000000000000000 t disk_change  [floppy]
0000000000000000 t set_dor      [floppy]
0000000000000000 t set_fdc      [floppy]
...

This cannot be used to build a proper mapping. Extend the test to
ensure we don't fail on the CI for such a case and also remove the
check on the found address, as it may point to 0 (see above). We
do want to get some name though.

Sometimes the kallsyms file only contains 0 addresses

Change-Id: I7f307ff89a9ec0cfd0f4d8e679e25a63db561bd0
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>