Default to the target system architecture
When parsing stream data, one used to be required to pass --arch,
even when trying to parse a perf.data file for the same architecture
as the current system. This patch simplifies this process a bit
by introducing a default architecture based on the target compiler
defines. This way, we don't need to pass --arch anymore when we
want to parse a perf data stream from the current machine.

Change-Id: I0aef6001cdd7f845754554ee2942546af4995e8e
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>