Do not invalidate caches when non-file-backed sections get overwritten
When we encounter ELF sections that are not file backed, such as //anon
or [heap], then we used to invalidate all caches. This is not required
as we never find any symbols in these sections anyways. Note that we
will still invalidate the cache, when a file-backed section is overwritten
by a non-file-backed section, or vice versa.

This patch greatly improves the performance of perfparser on my machine.
With a typical perf.data file of ~200MB size, I had to wait for ca. 33s.
Now, with this patch applied, I can parse the same file in ca. 9s.

Change-Id: Iad3905628308499636f5b13a2e542911521cce39
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>