Completely remove time handling from the PerfElfMap
Now that we sort the mmap and sample events by time, we can assume
that we will only look for mappings based on the current time, i.e.
the time of the last added elf mapping. This simplifies the code
a bit further and would allow us to optimize it later, if need be.

Note that this theoretically breaks the handling of samples that
violate the time ordering across FINISHED_ROUND events. This was
broken earlier already when we removed the overwritten elf mappings.
And note how this does not pose any real-world issues in my tests.

Change-Id: I24f14afdf17cf5d4f7dcb5440dc04d02f591fcb8
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>