Pass an explicit Elf* to deduce the target architecture
Instead of reporting a random elf map, open the first elf manually
and pass that to dwfl_attach_state. This will then be used to
deduce the target architecture and thus gives us more control about
what is going on, without influencing the actual mappings.

A perf record for a runtime-attached profile could potentially contain
mmap events for files first. This happened when attaching to KDevelop,
which opens files internally using mmap. In such cases, we would try
to guess the architecture from a text file containing code in text
form, instead of a binary ELF, which of course did not work.

Instead, validate the ELF we pass to dwfl_attach_state to guess the
architecture via elf_kind. If that returns ELF_K_NONE, the file
is not a valid ELF object and can thus not be used to guess the
architecture. We simply silently skip this and try with the next
elf file we encounter, until it hopefully works.

Change-Id: I00ec7fa1da669c4b5ed9156654818b64bdf050ef
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>