Include sample period and weight in parsed output
The sample period equals the performance counter value since the
last sample, and is e.g. emitted also by `perf script`.

The sample weight is important for client application to correctly
attribute the sampling cost. I.e. it is not enough to just count the
number of total sample, but rather one must use the weighted number
of samples.

Change-Id: I052ae25dcca972320ca8601b3d821398c08401ad
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>