Do not wait indefinitely when parsing empty/broken perf.data file
So far, perfparser hangs for cases like the following:

    echo -n > perf.data.empty
    ./perfparser --input perf.data.empty

Similarly, we can hang forever when we parse a file that only
contains a perf header but nothing else. This patch adds checks
on the file size for non-sequential inputs (i.e. files) which will
trigger an early return when the input data is broken.

Change-Id: I9c22010dd3628ef65e52a785e36c928445633570
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>