Send progress messages when parsing perf.data files
In these cases, we know the expected file size and can thus
compute our current progress value. We do this up to 100 times and
send the progress value as a normalized float percent, i.e. a value
between 0 and 1.

This is a helpful feature, as large data files take a long time
to parse. Showing the user that we make some progress is a good thing.

Change-Id: Icb0c9564e06173a526b726e93d75d4f5b7e8949d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>