Retry unwinding when the symbol cache is dirty
mmap()'d binaries can overlap in the profiled application's address
space. For example ld.so usually reserves a huge chunk of address space
for itself, which subsequently gets filled with other libraries.
However, libdw won't reliably accept binaries reported "on top of"
existing modules. In seemingly random cases it will complain that the
modules overlap and reject the new one.

So, when we fail to report a module to Dwfl, we mark the symbol cache
as dirty and retry unwinding at most once after clearing the cache.
Clearing the cache resets libdw's state and allows us to report the new
module first, and unwind symbols from it.

Change-Id: Idb5d85afb39e05c0439206b8d4938b79b6173b2c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>