Extend fallback code to find debug link file based on original path
By remembering the path under which a file was originally referenced
in the mmap call, we can use this information to find debug link
files. I.e. we will now try to find the debug link file in the
original binary target folder, which is where this debug link file
often resides.

This is useful for lazy people like me who don't want to set the
application path every time they use perfparser.

Change-Id: I8e33f137b186ed857b9b4da1d0618d0ae903d0cd
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>