Ensure we never overflow unsigned integers
When we loop over the unsigned integer and post-decrement it, the
very last iteration will lead to an overflow. This is harmless, as
we stop the iteration, but it still adds noise to UndefinedBehavior-
Sanitizer reports. This patch adds another check to remove the last
decrement to prevent this issue from arising.

Example warning:

3rdparty/perfparser/app/perfdata.cpp:491:22: runtime error: unsigned integer overflow: 0 - 1 cannot be represented in type 'quint64' (aka 'unsigned long long')
    #0 0x4d2aab in operator>>(QDataStream&, PerfRecordSample&) /home/milian/projects/kdab/rnd/hotspot/3rdparty/perfparser/app/perfdata.cpp:491:22

Change-Id: I1bd85b94a303dd73302cc86e0baa16a9fd835777