Load kallsyms from debug path buildid cache, if available
This allows us to parse perf.data files from other machines that
we distributed via `perf archive`. Additionally, we can now parse
older perf.data files from the same machine when the kernel got
updated in the meantime.

The kallsyms build-id cache is ignored when kallsyms is set
explicitly on the command line. Sometimes the kallsyms file in
the build-id cache is broken. Or maybe we just want to enforce
reading a different kallsyms file. This also mimicks the
behavior of `perf report --kallsyms` in such a case.

Change-Id: Iddec9ceb4e2e706305297c957ab85d29c247c9bb
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>