Always double-check the validity of the module known to libdwfl
Whenever we query for dwfl_addrmodule, we have to ensure the returned
module matches the mapping we expect. To do so, a new method is
introduced that wraps dwfl_addrmodule. If no module is found, or the
mapping mismatches, we report the expected elf map.

This does not fix any obvious issue but seems like the correct thing
to do. Additionally, it allows a follow-up fix.

Change-Id: I03208aaf6a64477ff7a8bdd5eb893ca34fd3038b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>