Set byte order on correct stream before parsing attribute IDs
Because the byte order was not set on the ID data stream, its values
where parsed in the wrong order, i.e. byte-swapped. This leads to
all kinds of errors later, when we try to lookup the attribute for
a given sample ID - where we parse the value in the correct byte
order.

Change-Id: I06dee77bfe2901daaf1d05c7a5079ae0194adf07
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>