Ignore the kallsyms build-id cache when kallsyms is set explicitly
Sometimes the kallsyms file in the build-id cache is broken. Or
maybe we just want to enforce reading a different kallsyms file.
Previously, this was impossible as the build-id always took
precedence over the kallsyms file passed on the command line.

Now, when the command line parameter is set explicitly, it will
be used and the build-id is ignored. This also mimicks the
behavior of `perf report --kallsyms` in such a case.

Change-Id: Ia5d529c915cedc6f39c662b4a69b22496441f982