Fix various warnings generated by clang
- pids and tids are generally signed on linux. Reflect that in the code.
- The counters in PerfUnwind::Stats need to be unsigned.
- All QIODevice offsets and read/write sizes need to be qint64.
- Register offsets are now signed int, to avoid excessive casting when
  using them to index QVector and QList.
- All the fixed struct lengths can be expressed in quin16.
- Replace C-style casts with static_cast.
- Add override to some method declarations.

As a consequence, a few cases where we did handle excessively long data
sections reveal themselves as unreachable, while in other places we have
to indeed take care of that. Instead of leaving the stream in an invalid
state, though, we generally skip those sections now. Traces of >2GB are
not unthinkable and who knows where that data shows up.

Change-Id: Ied5e5263fbe44b0635c2a6d0fc24fc511a4d0f54
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>