Report error only when accessing a missing elf file
This fixes an issue where we reported an error about missing elf files
for files that are not actually an elf file to begin with. E.g. I get
tons of errors similar to the following when profiling KDevelop:

  Could not find ELF file for /tmp/preamble-0abc76.pch.
  This can break stack unwinding and lead to missing symbols.

It happens when we encounter a file mapping for any file that got removed
in the meantime. Note that we get *all* mmap events, even for non-executable
file mappings. Reporting an error for those is confusing and unneccessary.

To improve this situation, instead only report an error when we actually
try to access a file that is unavailable.

Change-Id: I14df4de8c75dee5d5d3f61dbbba275bb3c78fa89