Report an error when we fail to find any elf for a given IP
This is also a serious issue that can completely break the analysis.
As such, report a user-visible error for that case. Note that the
IP address itself isn't too useful to users though. So that is not
part of the error message, to allow clients to only show this error
once.

Change-Id: I3883e5424de5ebf548a93e2cd563e51e84737b42