Cache address location information per elf file
Instead of using one large shared address cache that needs to be
cleared every now and then, cache the address location information
per elf file using relative addresses. This means we don't need to
clear this cache at all, as the lookup for a relative address into
a given elf file will always return the same data.

This greatly improves the performance under some situations where
the cache is cleared often. For one of my files, it drops the time
from 55s down to 18s.

Additionally, this patch paves the way to share (parts of) this cache
in PerfUnwind. Most of its contents are not PID specific anymore.

Change-Id: I79616fbb5c45a2543845df2d05d9936e49401627