Only invalidate symbol cache when a previously used elf gets invalidated
We previously cleared the cache whenever an mmap event overlapped
a previously encountered mmap event. But often, these mmaps events
where not (yet) used for dwfl. As such, we do not have anything cached
from them which would be invalidated now. So keep the rest of the cache
alive to speed up the process slightly for some scenarios. One of my
files now gets parsed in ~5s instead of ~6s before. I see that the
cache is now only cleared 500 times instead of 2500 times.

Change-Id: If9c29946f6fb12a6b6d64888fe34d8add8eaebb4