Speed up perfparser when DWARF ranges are broken/missing in ELFs
The previous approach to handle the broken DWARF emitter from clang
was to parse all CUs in the hope to find one that contains the
requested address. This was done repeatedly for every address and
often had to go through most of the CUs in a binary, which is a
costly process. To reproduce, one can do the following:

- build perfparser with clang
- profile this build of perfparser while parsing some given workload
- inspect the profile and notice the large overhead from
  find_fundie_by_pc

Instead, the first time dwfl_module_addrdie fails, we now query for
all CUs of an ELF and build a custom range map and use that for lookup.
The initial build of this range map is roughly as costly as querying
for two addresses using the old slow code path. As such, once we
query three or more addresses in a given ELF this new approach is
already yielding better performance. Some numbers from my test:

Before:
 Performance counter stats for './perfparser --input perf.data --output /dev/null':

      46576.925866      task-clock:u (msec)       #    0.999 CPUs utilized
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
            60,868      page-faults:u             #    0.001 M/sec
   144,506,727,842      cycles:u                  #    3.103 GHz
   472,211,018,102      instructions:u            #    3.27  insn per cycle
   116,423,898,539      branches:u                # 2499.605 M/sec
       488,663,345      branch-misses:u           #    0.42% of all branches

      46.611237448 seconds time elapsed

After:

 Performance counter stats for './perfparser --input perf.data --output /dev/null':

      17447.629837      task-clock:u (msec)       #    0.995 CPUs utilized
                 0      context-switches:u        #    0.000 K/sec
                 0      cpu-migrations:u          #    0.000 K/sec
            60,142      page-faults:u             #    0.003 M/sec
    53,150,847,387      cycles:u                  #    3.046 GHz
   158,503,454,039      instructions:u            #    2.98  insn per cycle
    38,711,860,671      branches:u                # 2218.746 M/sec
       190,209,418      branch-misses:u           #    0.49% of all branches

      17.543916999 seconds time elapsed

Change-Id: I9ca45ad7c8f77f91d0376f6dcae2f73c6e868404