Use dwfl_report_begin/_end correctly
The documentation states that you cannot use dwfl until dwfl_report_end
returns 0. In fact you can, but then you run into strange corner cases
where dwfl doesn't find certain modules that have been reported before.
So far it was assumed that we cannot report additional modules while
unwinding a stack, but in fact I don't see a reason why. So, whenever we
want to report an additional module we call dwfl_report_begin_add before
and dwfl_report_end afterwards. In addition, to make our dwfl initially
usable we do one round of dwfl_report_begin and dwfl_report_end when
creating it, and we have to change the order of begin and end on clear
because the usual state of the dwfl is now "closed for reporting".

Change-Id: I7468204c8f48a8a0c8f68f78e81f5c4aeb8fb8d4
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>