Make PerfStdin behave more like other QIODevices
Buffer the data and emit readyRead when new data is available. Avoid the
"oops" momemnt when bytesAvailable() returned >0 but trying to read the
bytes would instead close the device. In fact, that was the "normal" way
to finish the reading from stdin: We would create a garbage event header
by trying to read past the end of the input and then signal an error
because of that.

Now we can actually listen for aboutToClose() on PerfStdin and we will
eventually receive that signal.

Change-Id: Ib159d186f4ce6c69a1fc7fe37bd5673981307b68
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>