Try to unwind the user stack even if IP is in kernel space
Quite often we get both a kernel call chain in sample.callchain() and a
user stack and user registers. We should analyze both then, each with
the right symbol table.

Also, when resolving a call chain we need to make sure that every time
we resolve a frame the dwfl is attached. This is important as context
switches force us to change the symbol table.

Change-Id: I9e5849eefef6780f822166e8c6a798749d18fce2
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>