Don't repeatedly report unsuitable modules for an address
dwfl may disagree with our mmap events about the length of a module. If
dwfl considers the module to be shorter, the address we are looking for
may fall in an area that's not covered by the reported module. In that
case we should not try to report the same module again the next time we
catch this address. We already know that it doesn't fit, after all.

Task-number: QTCREATORBUG-20427
Change-Id: I852dd39f0a53fe7bc1bde149efa3541d0fc7b8ba
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>