If dwfl_module_getsrc fails, iterate CUs to find source file/line
Some ELF objects have .debug_info and .debug_line sections, but no
.debug_aranges. dwfl_module_getsrc doesn't like this and gives up. We
can, however, do better. The information is all there, we just need to
get hold of the right CU. The iteration is expensive, but as we cache
the resulting locations, it's tolerable.

In case there is no .debug_info at all, this won't iterate, as the Dwarf
object won't be found in the first place.

Task-number: QTCREATORBUG-20441
Change-Id: I9846397d1e930ba5f25ccfa9748542a198f0f2a4
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>