Use stable_sort instead of sort in PerfUnwind::flushEventBuffer
This is required to keep the order of events with the same time.
Most notably, this happens when we analyze data from a
runtime-attached perf record, i.e. `perf record -p ...`. There,
the data file will contain lots of mmap events with time 0. The
order is important though, so keep that via stable_sort.

Change-Id: Ibaa3a816b4921e404105f3d01f500f9ffb8ddd59