Try to guess offset base address before reporting elf to dwfl
When we record perf data files with `--call-graph [fp,lbr]`, we
only see mmap events for the executable code segments. These are
usually offset. Without the non-offset mmap event, we would not
report anything to dwfl so far. Now, we guess the base address and
use that before reporting the elf to dwfl. This seems to work, but
could fail once a gap would exist at the start.

Change-Id: I9d837e8bed650d6574e84401da67beeddaf7ff57
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>