Detect the presence of elfutils and don't build if it's unavailable
Just because the relevant files might be available from /usr we cannot
assume they actually are. Check for libdwfl.h and only built if that is
found. This also works on macOS in that it just skips the build unless
you have magically found a way to provide elfutils on macOS.

Change-Id: Ib402b43e3143a1c531a1d3f2a0be3480a08e0f82
Reviewed-by: Eike Ziller <eike.ziller@qt.io>