Validate base mapping before using it
Prevent infinite looping when we access a stale base map. This
could happen when we encounter bogus mmap lists as happens in
https://github.com/KDAB/hotspot/issues/164

Verify that the base map actually corresponds to the expected
elf map and only use that then. Otherwise don't use the base
map and continue with the original mapping, hoping for the best.

While this fixes the stack overflow of the initial bug report,
it doesn't solve the fundamental issue of dealing with broken
data... We'll have to figure that one out separately.

Fixes: https://github.com/KDAB/hotspot/issues/164

Change-Id: Iaebddbfbc891784a7fcc05df47aba761b75cc587