Simplify dwfl_module_addrdie fall-back when .debug_aranges is missing
We do not need to recurse the full DIE tree to build the mapping.
Instead, it's enough to find the CU DIE and its bias only. To do that,
we can simply look at the {low,high} PC value of the CU DIEs. If
either of these two values is missing, we look at the dwarf_ranges for
the CU DIE and use the min/max PC values of those to set the CU DIE
range.

In my tests this produces the same results as before. Note how the
inline frames etc. are later on found via dwarf_getscopes{,_die}.
Also, when we do have a .debug_aranges section, then the call to
dwfl_module_addrdie will return us also a CU DIE, not an inner DIE.

The simplified code is easier to understand, will consume less memory
and should also be faster to run.

This patch is based on feedback by David Blakie.

Change-Id: I97767e99b2957aa430b65589b32ec66a9479ff7d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>