Allow address cache lookup for ElfInfo with pgoff / baseAddr
I have encountered cases where the ElfInfo we pass to the address
cache is the pgoff-mmapped one. In that case, we need to take the
baseAddr to compute the relative address, not the absolute address
where the offsetted elf was mapped. In my case, the address was
actually outside of the pgoff-mmapped one and within a different
section apparently, which got handled properly by elfutils but our
caching didn't handle this yet - we just asserted and failed.

Change-Id: I2cd9d2cebbd60f00353ecbf413e020783374769e