Improve support for applications that rely on fork for parallelism
A forked process inherits the elfmap of the parent, thus we need to
copy that data when we encounter a fork event. This is complicated by
the fact that the task events and mmap events need to be sorted first.
Thus, we need to find the fork event when we deplete the buffers after
sorting and then initialize the child process elf map with the one
from the parent process.

Furthermore, pass the ppid through the ThreadStart event, to allow
client applications to inherit the comm for the newly created process.

Fixes: https://github.com/KDAB/hotspot/issues/241
Change-Id: I5de13644e12def6704c5f622428a815fd87d2af4