Free symbol table resources when a process exists
For perf.data files that span many (small) processes, we held on
to the symbol tables even after a process exited. This wastes
resources and can even lead to potential confusion when a PID is
recycled - we would have reused the symbol table from the last
process then, which could be a completely different one.

Change-Id: I858dce578619191babdd1a8410ad038c0b686ff6
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>