Reduce calls to dwarf_getscopes to improve performance
Implement an alternative approach to resolving inline frames and
cache subprogram die names.

The problem with dwarf_getscopes and dwarf_getscopes_die is that it
often has to traverse a large part of the DIE tree within a CU DIE
to find its result. For larger DSOs this repeated tree walking can
consume a lot of time.

The idea behind this patch is the following: We find (and cache)
the Dwarf_Die's with DW_TAG_subprogram and then look for inline frames
in that subtree directly.

Additionally, we cache the scope names within a CU DIE more
aggressively. I hope that in the future this area can be improved even
further, as the remaining calls to dwarf_getscopes_die still make up
a large fraction of the overall analysis cost.

The following shows the performance impact of this patch for a
perf.data file with about 6M samples. Many frames in the callstacks
point to a self-compiled libclang.so with debug symbols. That library
alone is roughly 600MB large. This makes finding inline frames quite
slow.

Before:

```
         80.159,75 msec task-clock                #    0,984 CPUs utilized
             4.075      context-switches          #    0,051 K/sec
                 1      cpu-migrations            #    0,000 K/sec
           152.257      page-faults               #    0,002 M/sec
   346.071.892.881      cycles                    #    4,317 GHz                      (83,33%)
     1.940.060.936      stalled-cycles-frontend   #    0,56% frontend cycles idle     (83,33%)
    38.399.679.774      stalled-cycles-backend    #   11,10% backend cycles idle      (83,34%)
   999.298.133.335      instructions              #    2,89  insn per cycle
                                                  #    0,04  stalled cycles per insn  (83,31%)
   239.561.868.424      branches                  # 2988,556 M/sec                    (83,34%)
     1.163.589.915      branch-misses             #    0,49% of all branches          (83,34%)

      81,497496973 seconds time elapsed

      79,554970000 seconds user
       0,404933000 seconds sys
```

After:

```
         15.558,09 msec task-clock                #    1,000 CPUs utilized
                99      context-switches          #    0,006 K/sec
                 1      cpu-migrations            #    0,000 K/sec
           151.446      page-faults               #    0,010 M/sec
    67.961.461.389      cycles                    #    4,368 GHz                      (83,32%)
       759.299.629      stalled-cycles-frontend   #    1,12% frontend cycles idle     (83,32%)
     7.369.116.441      stalled-cycles-backend    #   10,84% backend cycles idle      (83,34%)
   187.648.727.850      instructions              #    2,76  insn per cycle
                                                  #    0,04  stalled cycles per insn  (83,34%)
    45.231.315.052      branches                  # 2907,254 M/sec                    (83,34%)
       200.377.846      branch-misses             #    0,44% of all branches          (83,33%)

      15,560370834 seconds time elapsed

      15,230305000 seconds user
       0,285776000 seconds sys
```

Which means we are now roughly 5x faster than before, which is a
pretty significant gain.

Relates-To: https://github.com/KDAB/hotspot/issues/192
Change-Id: I0669cc3aad886b22165eaf1d0836a56e5183898d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>